+++ infra/bootstrapper/dev/locals.tf
@@ -1,0 +1,21 @@
+locals {
+  environment = {
+    prefix          = "dx"
+    location        = "italynorth"
+    location_short  = "itn"
+    env_short       = "d"
+    domain          = "devex"
+    app_name        = "core"
+    instance_number = "01"
+  }
+
+  tags = {
+    CostCenter     = "TS000 - Tecnologia e Servizi"
+    CreatedBy      = "Terraform"
+    Owner          = "DevEx"
+    Environment    = "Dev"
+    Source         = "https://github.com/pagopa/dx/blob/main/infra/bootstrapper/dev"
+    ManagementTeam = "Developer Experience"
+    TestName       = "Create Azure Github environment bootstrap for test"
+  }
+}