+++ infra/core/dev/outputs.tf
@@ -1,0 +1,3 @@
+output "values" {
+  value = module.core
+}