+++ infra/core/prod/locals.tf
@@ -1,0 +1,20 @@
+locals {
+  environment = {
+    prefix          = "dx"
+    location        = "italynorth"
+    location_short  = "itn"
+    env_short       = "p"
+    app_name        = "core"
+    instance_number = "01"
+  }
+
+  tags = {
+    CostCenter     = "TS000 - Tecnologia e Servizi"
+    CreatedBy      = "Terraform"
+    Owner          = "DevEx"
+    Environment    = "Prod"
+    Source         = "https://github.com/pagopa/dx/blob/main/infra/core/prod"
+    ManagementTeam = "Developer Experience"
+    TestName       = "Create Azure core resources for prod DEVEX environment"
+  }
+}