+++ infra/modules/azure_core_infra/_modules/github_runner/outputs.tf
@@ -1,13 +1,3 @@
-output "subnet_name" {
-  value       = azurerm_subnet.runner_snet.name
-  description = "Subnet name"
-}
-
-output "subnet_id" {
-  value       = azurerm_subnet.runner_snet.id
-  description = "Subnet id"
-}
-
 output "cae_id" {
   value       = azurerm_container_app_environment.cae.id
   description = "Container App Environment id"