+++ infra/modules/azure_core_infra/_modules/vpn/vpn.tf
@@ -1,15 +1,7 @@
 ## VPN
 
 data "azuread_application" "vpn_app" {
-  display_name = "eng-d-app-vpn"
-}
-
-resource "azurerm_subnet" "vpn_snet" {
-  name                 = "GatewaySubnet"
-  resource_group_name  = var.resource_group_name
-  virtual_network_name = var.virtual_network.name
-  address_prefixes     = [var.vpn_cidr_subnet]
-  service_endpoints    = []
+  display_name = "dx-${var.env_short}-app-vpn"
 }
 
 module "vpn" {
@@ -21,7 +13,7 @@ module "vpn" {
   sku                   = "VpnGw1"
   pip_sku               = "Standard"
   pip_allocation_method = "Static"
-  subnet_id             = azurerm_subnet.vpn_snet.id
+  subnet_id             = var.vpn_subnet_id
 
   vpn_client_configuration = [
     {
@@ -40,29 +32,13 @@ module "vpn" {
   tags = var.tags
 }
 
-## DNS FORWARDER
-resource "azurerm_subnet" "dns_forwarder_snet" {
-  name                 = "${var.project}-dns-forwarder-snet-01"
-  resource_group_name  = var.resource_group_name
-  virtual_network_name = var.virtual_network.name
-  address_prefixes     = [var.dnsforwarder_cidr_subnet]
-
-  delegation {
-    name = "delegation"
-
-    service_delegation {
-      name    = "Microsoft.ContainerInstance/containerGroups"
-      actions = ["Microsoft.Network/virtualNetworks/subnets/action"]
-    }
-  }
-}
-
+## DNS FORWARDR
 module "dns_forwarder" {
   source              = "github.com/pagopa/terraform-azurerm-v4//dns_forwarder?ref=v1.9.0"
   name                = "${var.project}-dns-forwarder-ci-01"
   location            = var.location
   resource_group_name = var.resource_group_name
-  subnet_id           = azurerm_subnet.dns_forwarder_snet.id
+  subnet_id           = var.dnsforwarder_subnet_id
 
   tags = var.tags
 }